Move ModuleBuilder to src only, test now imports from src
